Added support for environment variables
This adds support for the `REACT_APP_GITHUB_TOKEN` and the (optional) `REACT_APP_GITHUB_NAME`.
This commit aims to protect private keys, since (as little privilege as they may have) it is a security risk to expose any such private keys. It is much safer to set them in environment variables, where they stay on the local machine restricted to that specific user.